Document and cleanup specimen options
I found that the only way to verify the usage of these options was to build and document it. It turned out that spaces are not allowed as language parameters, which is why I used the format "style|option1,option2". Let's see where to go from here regarding passing CSS in.

References #5