Clone styles from parent document into iframe
This is a fix for components which use webpack's style-loader. style-loader injects the styles directly into the document's head when the component is loaded. In order to get styled components in framed React and HTML specimens, these styles need to be cloned into the generated iframe.

This doesn't clone any Catalog styles because they are either inline styles or part of the body.