Publish builds to gh-pages
We now publish builds of Catalog to gh-pages again.

`make dist` also got a bit smarter. It now determines a prerelease on the version number in package.json (if it contains a dash, it's a prerelease). Use it to bump the version, publish to npm, and to publish to gh-pages.

Every version gets its own directory on gh-pages including the full documentation for this version and the built files. Additionally, stable releases get published to the top level and prereleases to `next/`.

Fixes #90