Refactor Catalog context
We now define only one `catalog` context object (which contains all previously defined context props). This makes Catalog a better citizen because it doesn't pollute context anymore and it's clearer which context props belong to which library.

Specimens now don't have to use context at all because the current catalog context object will be passed as a prop to them. Before they just received `theme` as a prop which wasn't enough for `ReactSpecimen`. This opens the door for more advanced specimens and also makes them easier to test (once we get to that).