Fix responsive frame layout
In some edge cases, the iframe got scaled twice (e.g. one `transform: scale(0.5)` somehow got applied twice, resulting in an iframe with scale 0.25). Wrapping it in an element with the scale transform *and* defined width and height solved this issue.

To prevent a rendering glitch, the React and HTML specimens now only show the frame after their own width is measured.