Responsive mode layout fix (#264)
* Fix responsive frame layout

In some edge cases, the iframe got scaled wrong (e.g. one `transform: scale(0.5)` somehow got applied twice, resulting in an iframe with scale 0.25). Wrapping it in an element with the scale transform *and* defined width and height solved this issue.

To prevent a rendering glitch, the React and HTML specimens now only show the frame after their own width is measured.

* Fix horizontal scroll bars

Not sure why this ever worked with 100% width and a margin ¯\_(ツ)_/¯