Allow expanding the source via prop
Fixes #356

Add `showSource` prop to html specimen to allow rendering with the source already expanded.

Apply this to the react specimen as well – there is no apparent reason why they should behave differently.

For multiple columns, the specimen content now fills the available height while the the source is placed at the bottom instead of creating a gap.