switch from em to rem
When a table specimen is used with a `baseFontSize` other than tha default `16px` the table cells are displayed with a strange font size. That's because the font size of the `table` gets multiplied with the one of the `cell` or the `paragraph`. For example for `baseFontSize` set to `18px`: `18 / 16 = 1.125em` (which will result in a calculated font size of `22.7833px` (`1.25 * 1.25 * 18`).

Switching to rem (root em) solves this issue.