Typesetting improvements (#446)
* switch from em to rem

When a table specimen is used with a `baseFontSize` other than tha default `16px` the table cells are displayed with a strange font size. That's because the font size of the `table` gets multiplied with the one of the `cell` or the `paragraph`. For example for `baseFontSize` set to `18px`: `18 / 16 = 1.125em` (which will result in a calculated font size of `22.7833px` (`1.25 * 1.25 * 18`).

Switching to rem (root em) solves this issue.

* Fix elements nested in blockquotes

* Add more variations to test page

* Improve headings in hints

* Remove unnecessary text styles on <table>

* Add comment about blockquote style

* Define letter spacing

* Remove unused sidebar theming option

* More specific styling of sidebar

Doesn't solve #347 completely but addresses some potential issues with global stylesheets

* More spefic list styling in content

* More specific link styling in content

* More specific navigation bar styling

Also changes unsemantic headings to divs, fixes #444

* More specific link styling in Download specimen

Also removes non-semantic headings, fixes #444

* More layout fixes

* Fix nested list margin

* Text fixes in color specimens

* Fix responsive tabs text styles

* Fix hint link styles

* Update test snapshots

* Make blockquoteStyle lazy to fix font loading

`@import`s in `insertRule` only work if they're the first rule
https://developer.mozilla.org/en-US/docs/Web/API/CSSStyleSheet/insertRule#Restrictions

* Upgrade create-emotion

* Fix single active link padding